fixed:blur logo and active task card UI
