+++ package.json
@@ -55,7 +55,7 @@
     "react-native-reanimated": "^2.2.4",
     "react-native-safe-area-context": "^3.2.0",
     "react-native-screens": "^3.9.0",
-    "react-native-svg": "^14.0.0",
+    "react-native-svg-uri": "^1.2.3",
     "react-native-tab-view": "^3.5.2",
     "react-native-toast-message": "^2.1.5",
     "react-native-walkthrough-tooltip": "^1.5.0",
