+++ src/screens/ProfileScreen/ProfileScreen2.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useCallback, useContext } from 'react';
-import { View, ScrollView, StyleSheet } from 'react-native';
+import { View, TouchableWithoutFeedback } from 'react-native';
 import { ScreenViewContainer } from '../../styles/GlobalStyle';
 import { profileScreenStyles } from './styles';
 import Avatar from '../../components/Avatar';
@@ -8,49 +8,10 @@ import { AuthContext } from '../../context/AuthContext';
 import { ImagePickerResponse } from 'react-native-image-picker';
 import All from './TaskScreens/All';
 import { Tabs } from 'react-native-collapsible-tab-view';
-
-import { useFocusEffect } from '@react-navigation/native';
-import { fetchActiveTasks } from '../AuthScreen/Util';
-import DisplayContribution from '../../components/DisplayContribution';
 import UserData from './User Data/UserData';
-import Loader from '../../components/Loader';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import EllipseComponent from '../../components/EllipseComponent';
-
-
-export const ActiveScreen = () => {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [activeTasks, setActiveTasks] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const { loggedInUserData } = useContext(AuthContext);
-
-  useFocusEffect(
-    useCallback(() => {
-      setLoading(true);
-      (async () => {
-        const token = loggedInUserData?.token;
-
-        const tasksRes = await fetchActiveTasks(token);
-        const activeTaskRes = tasksRes.filter(
-          (item) => item.status !== 'COMPLETED',
-        );
-        setActiveTasks(activeTaskRes);
-        setLoading(false);
-      })();
-
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [loggedInUserData?.token]),
-  );
-  return (
-    <View style={styles.profile}>
-      {loading ? (
-        <Loader />
-      ) : (
-        <DisplayContribution tasks={activeTasks} expand={false} />
-      )}
-    </View>
-  );
-};
+import ActiveScreen from './TaskScreens/ActiveTask';
 
 const ProfileScreen = () => {
   // const dispatch = useDispatch();
@@ -58,6 +19,11 @@ const ProfileScreen = () => {
   const [response, setResponse] = useState<ImagePickerResponse>({});
   const [modalVisible, setModalVisible] = useState(false);
   const { loggedInUserData, setLoggedInUserData } = useContext(AuthContext);
+  const [isDropdownVisible, setDropdownVisible] = useState(false);
+
+  const handleDropdown = () => {
+    setDropdownVisible((prev) => !prev);
+  };
 
   const closeModal = useCallback(() => {
     setModalVisible(false);
@@ -81,14 +47,15 @@ const ProfileScreen = () => {
   };
 
   return (
-    <ScrollView contentContainerStyle={ScreenViewContainer.container}>
-      {/* <Pressable
-        style={profileScreenStyles.logoutButton}
-        onPress={handleLogout}
-      >
-        <Text style={profileScreenStyles.logoutText}>{Strings.LOGOUT}</Text>
-      </Pressable> */}
-      <EllipseComponent handleLogout={handleLogout} />
+    <TouchableWithoutFeedback
+      contentContainerStyle={ScreenViewContainer.container}
+      onPress={handleDropdown}
+    >
+      <EllipseComponent
+        handleLogout={handleLogout}
+        isDropdownVisible={isDropdownVisible}
+        handleDropDown={handleDropdown}
+      />
       <UploadImageModalView
         closeModal={closeModal}
         modalVisible={modalVisible}
@@ -118,7 +85,7 @@ const ProfileScreen = () => {
           }}
         /> */}
       </View>
-    </ScrollView>
+    </TouchableWithoutFeedback>
   );
 };
 
@@ -138,12 +105,5 @@ const ProfileScreen2: React.FC = ({ navigation }) => {
     </Tabs.Container>
   );
 };
-
-const styles = StyleSheet.create({
-  profile: {
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-});
 
 export default ProfileScreen2;
