+++ src/screens/ProfileScreen/TaskScreens/ActiveTask.tsx
@@ -1,41 +1,50 @@
-import React, { useCallback, useContext, useState } from 'react';
+import React, { useState, useContext, useCallback } from 'react';
+import { View, StyleSheet } from 'react-native';
 import { useFocusEffect } from '@react-navigation/native';
-import { fetchContribution } from '../../AuthScreen/Util';
 import { AuthContext } from '../../../context/AuthContext';
 import DisplayContribution from '../../../components/DisplayContribution';
-import { View, StyleSheet } from 'react-native';
+import Loader from '../../../components/Loader';
+import { fetchActiveTasks } from '../../AuthScreen/Util';
 
 const ActiveScreen = () => {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [activeTasks, setActiveTasks] = useState([]);
+  const [loading, setLoading] = useState(false);
   const { loggedInUserData } = useContext(AuthContext);
 
   useFocusEffect(
     useCallback(() => {
+      setLoading(true);
       (async () => {
-        const userName = loggedInUserData?.username;
-        const contributionResponse = await fetchContribution(userName);
-        setActiveTasks(
-          contributionResponse.all.filter(
-            (item: { task: { status: string } }) =>
-              item.task.status === 'ACTIVE',
-          ),
+        const token = loggedInUserData?.token;
+
+        const tasksRes = await fetchActiveTasks(token);
+        const activeTaskRes = tasksRes.filter(
+          (item) => item.status !== 'COMPLETED',
         );
+        setActiveTasks(activeTaskRes);
+        setLoading(false);
       })();
+
       // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []),
+    }, [loggedInUserData?.token]),
   );
   return (
     <View style={styles.profile}>
-      <DisplayContribution tasks={activeTasks} expand={true} />
+      {loading ? (
+        <Loader />
+      ) : (
+        <DisplayContribution tasks={activeTasks} expand={false} />
+      )}
     </View>
   );
 };
 
+export default ActiveScreen;
+
 const styles = StyleSheet.create({
   profile: {
-    justifyContent: 'flex-start',
-    alignItems: 'center',
+    // justifyContent: 'center',
+    // alignItems: 'center',
   },
 });
-
-export default ActiveScreen;
